style to match wireframe
