+++ client/app/login/page.tsx
@@ -1,8 +1,13 @@
 import LoginForm from '@components/LoginAndRegisterHandler/Login/LoginForm'
+import ChameleonGraphic from '@components/ChameleonGraphic/ChameleonGraphic'
+import IntroText from '@components/IntroText/IntroText'
+import styles from '@styles/Register.module.scss'
 
 const Login = () => {
   return (
-    <div>
+    <div className={styles.pageLayout}>
+      <ChameleonGraphic />
+      <IntroText />
       <LoginForm />
     </div>
   )
