+++ client/app/register/layout.tsx
@@ -1,0 +1,7 @@
+interface Props {
+  children: React.ReactNode
+}
+
+export default function Layout({ children }: Props) {
+  return <div>{children}</div>
+}
