+++ client/components/IntroText/introText.module.scss
@@ -5,6 +5,6 @@
     flex-direction: column;
     align-items: center;
     color: v.$neutral-500;
-    font-size: v.$font-size-lg;
+    font-size: v.$base-font-size * 1.1;
     gap: .5rem;
 }
