--- client/components/LoginAndRegisterHandler/Register/Register.tsx
@@ -1,9 +1,0 @@
-const Register = () => {
-  return (
-    <div style={{ height: '100px', width: '100px', backgroundColor: 'red' }}>
-      test
-    </div>
-  )
-}
-
-export default Register
