+++ client/components/LoginAndRegisterHandler/Register/RegisterForm.module.scss
@@ -80,7 +80,7 @@
     flex-basis: 45%;
 }
 
-.loginButton {
+.submitButton {
     background-color: variables.$primary-dark;
     color: variables.$neutral-200;
     flex-basis: 45%;
