+++ client/styles/Register.module.scss
@@ -1,9 +1,7 @@
-@use "variables.module.scss" as v;
-
-.homeLayout {
+.pageLayout {
     display: flex;
     flex-direction: column;
     align-items: center;
-    gap: 3rem;
+    gap: 2rem;
     margin-top: 2rem;
 }
